Improve reliability of benchmarking scripts.
On my Raspberry Pi I often ran into the problem that the Pi would miss the
beginning of the serial output and then get stuck in an infinite loop trying
to reflash again and again.
By waiting a couple of hundred ms when starting up, this can be prevented.
For me this heavily improved reliability.